Refactor calcNewValue
Previously, calcNewValue assumed your display's maxValue was a 100.
This caused the OSD to block after certain values (see #201).
It also tried to find the "prettiest" monitor value to display during
normal increments.

This PR refactors the calcNewValue function to focus on finding the next value based on the monitor's value scale.
Behaviour:
- Get the corresponding chiclet value based on the monitor value.
- Calculate the distance between the nearest whole chiclet and the chiclet
from the monitor value
- Depending on the direction (up or down) and if the distance is between
a set threshold, we jump to the next chiclet skipping the nearest.
Otherwise, we go to the nearest chiclet.

calcNewValue no longer tries to find the "prettiest" value as we round the chiclets during normal increments and use the chiclet scale (x/16).